requirements.txt
Django==5.0.6
djangorestframework==3.15.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==22.0.0
